Add keyring lock-identity operation
Whenever a single identity is locked (released), all other identities
with the same PIN are flagged to indicate that the PIN is not "fully"
unlocked, so that the next time the same PIN is entered, the slot
decryption is re-tried for non-loaded identities, and the locked
identity will be unlocked again.

Added a test for the GET /restful/keyring/SID/lock request.